Install OSX command line tools
Summary of Changes
==================

Since the OS X Command Line Tools are a hard dependency for Homebrew, this
role will now make sure that they are installed. If they are already
installed, those steps will be skipped. This makes using this role easier
to use when provisioning MacOS, especially when the destination host is not
the same as the control node.

Other related changes:

* `meta/main.yml`: specifies platform `MacOSX` instead of `GenericUNIX`
* `.travis.yml`: Fix homebrew uninstall process-- the old way hangs.
* `tasks/main.yml`: Ensure `root` is the owner of `homebrew_prefix` (`/usr/local`)