Enhancement - Clear Cache
Summary of Changes
==================

This change provides the option to clear the Homebrew cache after any software
installation. The handler will only run if the related role variable is set.

Why?
----

The benefit is that this saves disk space and can avoid installation/execution
of potentially corrupt software, but the cost is the time saved by avoiding
re-downloading what were otherwise cached files (fixes #40).