Use loop instead of with_items
Looping syntax is changing and so this change updates how looping
is peformed for future compatibility.

Since the `squash_actions` behavior in `with_X` loops has been
deprecated and will be removed in Ansible 2.11, there continues to
be noisy deprecation warnings. This persists even when attempting to
override the `squash_actions` setting to remove the `homebrew` module
from the [default list][default-squash-actions].

Fixes #110

[default-squash-actions]: https://docs.ansible.com/ansible/2.7/reference_appendices/config.html?highlight=squash_actions#default-squash-actions