Add homebrew_user and homebrew_group variables (#111)
New Description
===============

This PR adds new variables `homebrew_user` and `homebrew_group` to enable users to specify an alternative user (and Unix group) when `ansible_user_id` is insufficient. See the conversation in the PR for details.

Original Description
====================

If the default install path of `/usr/local` is used, then the user creating directories within that path must be root. 

My use case is that I create a `buildkite` user for use on our CI server as part of my role. The user does not exist at the start of my role, and I expect at the end of the process that the user will be unprivileged. 

I'd been doing the following to get this to work:

```yaml
  roles:
    - role: geerlingguy.homebrew
      ansible_user_id: buildkite
      become: yes
      become_user: buildkite
```

Maybe I'm doing this wrong?