DRY up privilege escalation settings (#113)
This is a follow up for PR #111.

To avoid repetitious settings, this change moves several tasks
with identical privilege escalation settings (i.e. `become*`) into
a `block` structure.

Furthermore, privilege escalation need only be invoked if the
`homebrew_user` and `ansible_user_id` are mismatched, so this change
avoids unnecessary requests for privilege (e.g. `sudo -u`). In fact,
this helps preserve the original behavior prior to #111.

Finally, this change updates the handlers to run as the intended user.